code cleanup on topnav and app, refactor styles to sass
