+++ src/App.tsx
@@ -1,23 +1,17 @@
 import React, { useState, useEffect } from 'react'
-import './styles/style.scss'
 import {
   BrowserRouter as Router,
   Switch,
   Route,
   Redirect,
 } from 'react-router-dom'
+import { makeStyles } from '@material-ui/core/styles'
+import CssBaseline from '@material-ui/core/CssBaseline/CssBaseline'
+import { ThemeProvider, Typography, Grid } from '@material-ui/core'
 import Intro from './components/static/Intro'
 import EligibilityRegistration from './components/registration/EligibilityRegistration'
 import Login from './components/login/Login'
 import { useSessionDataState, useSessionDataDispatch } from './AuthContext'
-import { makeStyles } from '@material-ui/core/styles'
-import CssBaseline from '@material-ui/core/CssBaseline/CssBaseline'
-import {
-  createMuiTheme,
-  ThemeProvider,
-  Typography,
-  Grid,
-} from '@material-ui/core'
 import { getSearchParams } from './helpers/utility'
 import Dashboard from './components/dashboard/Dashboard'
 import { TopNav } from './components/widgets/TopNav'
@@ -28,127 +22,8 @@ import Footer from './components/widgets/Footer'
 import PrivacyPolicy from './components/static/PrivacyPolicy'
 import { UserDataGroup, SessionData } from './types/types'
 import { ElegibilityProvider } from './components/registration/context/ElegibilityContext'
-
-const fallbackFonts = [
-  'OpenSans',
-  'serif',
-  'Roboto',
-  'Helvetica',
-  'Arial',
-  'Raleway',
-  'Quicksand',
-]
-
-export const systemFonts = ['Lato', ...fallbackFonts].join(',')
-
-export const playfairDisplayFont = ['Playfair Display', ...fallbackFonts].join(
-  ',',
-)
-
-export const sourceSansFont = ['Source Sans Pro', ...fallbackFonts].join(',')
-
-const defaultTheme = createMuiTheme()
-const theme = createMuiTheme({
-  typography: {
-    // Tell Material-UI what's the font-size on the html element is.
-    htmlFontSize: 10,
-    fontFamily: systemFonts,
-    button: {
-      textTransform: 'none',
-    },
-  },
-  palette: {
-    background: {
-      default: '#f5f5f5',
-    },
-    primary: {
-      // light: will be calculated from palette.primary.main,
-      main: '#343F56', ///'#202423' //'#ff4400',
-      // dark: will be calculated from palette.primary.main,
-      // contrastText: will be calculated to contrast with palette.primary.main
-    },
-    secondary: {
-      //light: '#0066ff',
-      main: '#ccc',
-      // dark: will be calculated from palette.secondary.main,
-      //contrastText: '#ffcc00',
-    },
-    // Used by `getContrastText()` to maximize the contrast between
-    // the background and the text.
-    contrastThreshold: 3,
-    // Used by the functions below to shift a color's luminance by approximately
-    // two indexes within its tonal palette.
-    // E.g., shift from Red 500 to Red 300 or Red 700.
-    tonalOffset: 0.2,
-  },
-  props: {
-    // Name of the component ⚛️
-    MuiButtonBase: {
-      // The properties to apply
-      disableRipple: true, // No more ripple, on the whole application 💣!
-    },
-  },
-  overrides: {
-    MuiOutlinedInput: {
-      root: {
-        borderRadius: 0,
-      },
-    },
-    MuiSwitch: {
-      track: {
-        backgroundColor: '#807C7C',
-      },
-    },
-    MuiButton: {
-      root: {
-        borderRadius: 25,
-        height: 47,
-        fontFamily: systemFonts,
-        fontWeight: 700,
-        color: 'white',
-      },
-      text: {
-        borderRadius: 25,
-        height: 47,
-        fontFamily: systemFonts,
-        color: 'white',
-        '&:hover': {
-          background: 'none',
-          textDecoration: 'underline',
-        },
-      },
-      containedPrimary: {
-        color: 'white',
-      },
-    },
-    MuiInputBase: {
-      root: {
-        fontFamily: systemFonts,
-      },
-    },
-    MuiCard: {
-      root: {
-        backgroundColor: 'white',
-        maxWidth: '511px',
-        margin: '0 auto',
-        padding: '40px',
-      },
-    },
-    MuiCardContent: {
-      root: {
-        [defaultTheme.breakpoints.up('md')]: {
-          padding: '46px',
-        },
-
-        '&:last-child': {
-          [defaultTheme.breakpoints.up('md')]: {
-            paddingBottom: '46px',
-          },
-        },
-      },
-    },
-  },
-})
+import { theme } from './theme'
+import './styles/style.scss'
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -253,31 +128,6 @@ function App() {
 
   const classes = useStyles()
 
-  const getTopClass = (location: string) => {
-    const alertClass = !!sessionData.alert ? ' hasAlert' : ''
-    const specialPages = ['survey', 'contactinfo', 'appointment']
-    if (specialPages.find(page => location.toLowerCase().includes(page))) {
-      return `partialGreen${alertClass}`
-    }
-    //dashboard is green for users to haven't tested
-    if (
-      location.toLowerCase().includes('dashboard') &&
-      sessionData.userDataGroup.indexOf('tests_available') === -1
-    ) {
-      return `partialGreen${alertClass}`
-    } else {
-      return ''
-    }
-  }
-
-  const shouldShowFooter = (location: string): boolean => {
-    const specialPages = ['dashboard', 'survey', 'contactinfo', 'consent']
-    return (
-      specialPages.find(page => location.toLowerCase().includes(page)) ===
-      undefined
-    )
-  }
-
   return (
     <ThemeProvider theme={theme}>
       <ElegibilityProvider>
@@ -285,7 +135,7 @@ function App() {
           <div className={classes.root}>
             <CssBaseline />
             <Router>
-              <div className={getTopClass(currentLocation)}>
+              <div>
                 <GoogleAnalyticsPageTracker />
                 <ScrollToTopOnRouteChange
                   onRouteChangeFn={(location: string) =>
@@ -347,7 +197,7 @@ function App() {
                     </Route>
                   </Switch>
                 </TopNav>
-                {shouldShowFooter(currentLocation) && <Footer token={token} />}
+                <Footer token={token} />
               </div>
             </Router>
           </div>
