+++ src/styles/style.scss
@@ -11,6 +11,7 @@
 @import 'components/sage-form';
 @import 'material-overrides';
 @import 'forms';
+@import 'components/topnav';
 @import 'mobile' screen and (max-width: 768px);
 
 html {
